pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2"
python-dotenv = "^1.0.1"
redis = "^5.1.1"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
pytils = "^0.4.1"
networkx = "^3.4.2"
matplotlib = "^3.9.2"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.4"
coreapi = "^2.3.3"
setuptools = "^75.3.0"
django-cors-headers = "^4.6.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
